new: testable version
