+++ .gitignore
@@ -1,5 +1,6 @@
 # Logs
 logs
+dist
 *.log
 npm-debug.log*
 package-lock.json
