+++ lib/index.ts
@@ -1,7 +1,7 @@
 //
 //
 // export API
-export { Concepts } from './concepts';
+//export { Concepts } from './concepts';
 export { Customers } from './customers';
-export { Machine } from './machine';
+export { MachineCreate } from './machine-create';
 export { MachineIndex } from './machine-index';
