--- test/data/orders-subset.json
@@ -1,355 +1,0 @@
-[
-  {
-    "oid": 2001703,
-    "shipping": {
-      "bags": 2,
-      "postalCode": "1203",
-      "when": "2017-05-12T14:00:00.000Z"
-    },
-    "customer": {
-      "id": 15,
-      "pseudo": "E**L",
-      "created": {
-        "$date": "2016-11-05T12:50:36.644Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 2,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c2"
-      }
-    ]
-  },
-  {
-    "oid": 2000808,
-    "shipping": {
-      "postalCode": "1208",
-      "when": "2016-05-03T14:00:00.000Z"
-    },
-    "customer": {
-      "id": 15,
-      "pseudo": "F**N",
-      "created": {
-        "$date": "2014-12-10T12:45:47.195Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 4,
-        "vendor": "d",
-        "quantity": 2,
-        "category": "c3"
-      }
-    ]
-  },
-  {
-    "oid": 2000437,
-    "shipping": {
-      "bags": 1,
-      "postalCode": "1205",
-      "when": "2015-11-27T15:00:00.000Z"
-    },
-    "customer": {
-      "id": 15,
-      "pseudo": "K**J",
-      "created": {
-        "$date": "2015-11-23T22:47:26.324Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 2,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c2"
-      },
-      {
-        "sku": 3,
-        "vendor": "c",
-        "quantity": 1,
-        "category": "c2"
-      },
-      {
-        "sku": 4,
-        "vendor": "d",
-        "quantity": 2,
-        "category": "c3"
-      }
-    ]
-  },
-  {
-    "oid": 2000859,
-    "shipping": {
-      "bags": 2,
-      "postalCode": "1208",
-      "when": "2016-05-20T14:00:00.000Z"
-    },
-    "customer": {
-      "id": 15,
-      "pseudo": "C**D",
-      "created": {
-        "$date": "2014-11-19T12:42:16.093Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 2,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c2"
-      },
-      {
-        "sku": 3,
-        "vendor": "c",
-        "quantity": 1,
-        "category": "c2"
-      }
-    ]
-  },
-  {
-    "oid": 2000740,
-    "shipping": {
-      "bags": 1,
-      "postalCode": "1208",
-      "when": "2016-04-05T14:00:00.000Z"
-    },
-    "customer": {
-      "id": 5,
-      "pseudo": "C**D",
-      "created": {
-        "$date": "2014-11-19T12:42:16.093Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 4,
-        "vendor": "d",
-        "quantity": 2,
-        "category": "c3"
-      },
-      {
-        "sku": 5,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 6,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c1"
-      }
-    ]
-  },
-  {
-    "oid": 2001939,
-    "shipping": {
-      "postalCode": "1203",
-      "when": "2017-08-22T14:00:00.000Z"
-    },
-    "customer": {
-      "id": 5,
-      "pseudo": "B**L",
-      "created": {
-        "$date": "2016-02-25T10:37:58.257Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 5,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 6,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c1"
-      }
-    ]
-  },
-  {
-    "oid": 2000539,
-    "shipping": {
-      "bags": 1,
-      "postalCode": "1208",
-      "when": "2016-01-12T15:00:00.000Z"
-    },
-    "customer": {
-      "id": 5,
-      "pseudo": "F**N",
-      "created": {
-        "$date": "2014-12-10T12:45:47.195Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 3,
-        "vendor": "c",
-        "quantity": 1,
-        "category": "c2"
-      },
-      {
-        "sku": 4,
-        "vendor": "d",
-        "quantity": 2,
-        "category": "c3"
-      },
-      {
-        "sku": 5,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 6,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c1"
-      }
-    ]
-  },
-  {
-    "oid": 2000482,
-    "shipping": {
-      "bags": 3,
-      "postalCode": "1207",
-      "when": "2015-12-08T15:00:00.000Z"
-    },
-    "customer": {
-      "id": 5,
-      "pseudo": "F**M",
-      "created": {
-        "$date": "2015-09-21T12:02:31.026Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 2,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c2"
-      },
-      {
-        "sku": 5,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      }
-    ]
-
-  },
-  {
-    "oid": 2000590,
-    "shipping": {
-      "bags": 1,
-      "postalCode": "1227",
-      "when": "2016-01-29T15:00:00.000Z"
-    },
-    "customer": {
-      "id": 5,
-      "pseudo": "F**C",
-      "created": {
-        "$date": "2014-12-13T10:50:39.913Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 5,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 6,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c1"
-      }
-    ]
-
-  },
-  {
-    "oid": 2000559,
-    "shipping": {
-      "postalCode": "1227",
-      "when": "2016-01-19T15:00:00.000Z"
-    },
-    "customer": {
-      "id": 5,
-      "pseudo": "F**C",
-      "created": {
-        "$date": "2014-12-13T10:50:39.913Z"
-      },
-      "likes": []
-    },
-    "items": [
-      {
-        "sku": 1,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 5,
-        "vendor": "a",
-        "quantity": 1,
-        "category": "c1"
-      },
-      {
-        "sku": 6,
-        "vendor": "b",
-        "quantity": 1,
-        "category": "c1"
-      }
-    ]
-  }
-]
