--- test/data/products-subset.json
@@ -1,93 +1,0 @@
-[
-  {
-    "sku": 1,
-    "created": "2014-11-19T19:38:13.046Z",
-    "updated": "2018-06-19T09:46:47.786Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c1",
-    "vendor": "a",
-    "natural": true,
-    "local": true
-  },
-  {
-    "sku": 2,
-    "created": "2014-11-20T13:21:56.415Z",
-    "updated": "2018-06-19T09:46:11.166Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c2",
-    "vendor": "b",
-    "natural": true,
-    "local": true
-  },
-  {
-    "sku": 3,
-    "created": "2014-11-20T14:24:18.871Z",
-    "updated": "2018-06-19T09:46:49.544Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c2",
-    "vendor": "c",
-    "natural": true,
-    "local": true
-  },
-  {
-    "sku": 4,
-    "created": "2014-11-20T14:44:03.035Z",
-    "updated": "2018-06-19T09:45:57.119Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c3",
-    "vendor": "d",
-    "natural": true,
-    "local": true
-  },
-  {
-    "sku": 5,
-    "created": "2014-11-20T14:46:14.094Z",
-    "updated": "2019-10-07T13:41:20.348Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c1",
-    "vendor": "a",
-    "natural": true,
-    "local": true
-  },
-  {
-    "sku": 6,
-    "created": "2014-11-20T14:53:29.716Z",
-    "updated": "2019-11-14T12:14:55.339Z",
-    "discount": false,
-    "boost": true,
-    "categories": "c1",
-    "vendor": "b",
-    "natural": false,
-    "local": true
-  },
-  {
-    "sku": 7,
-    "created": "2014-11-23T10:42:31.114Z",
-    "updated": "2018-06-19T09:51:45.405Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c3",
-    "vendor": "a",
-    "natural": true,
-    "local": true
-  },
-  {
-    "sku": 8,
-    "title": "Beurre en motte",
-    "slug": "beurre-en-motte",
-    "photo": "//ucarecdn.com/90bb2376-7024-4cfd-8671-1258b293d946/",
-    "created": "2014-11-23T19:49:59.620Z",
-    "updated": "2019-11-01T19:09:53.777Z",
-    "discount": false,
-    "boost": false,
-    "categories": "c3",
-    "vendor": "a",
-    "natural": true,
-    "local": false
-  }
-]
